Fix failed `yarn install` by ensuring lockfile producer and consumer on same version
The bug:

- Because yarn pnp is a giant hack, it needs to patch TypeScript for compatibilitiy,
   which it does by rewriting the TypeScript identifier to use a specific patch that `yarn`
   creates after each TypeScript release (aside: this would not normally affect us, since
   we use the `node-modules` linker instead of pnp. For us, applying the patch is a bug,
   and will not happen in `3.0.0`)
- There was a bug in this patch that was fixed in `v2.4.2`, so the hash of the patch
   is different depending on the version of `yarn` that is installing it (note: this is likely
   an example of why yarn wants us to check in the exact yarn source)
- Every time `./setup.sh` ran, it installed the latest yarn version (currently `2.4.2`)
- The lockfile was generated by `v2.4.1` but being consumed in CI by the (freshly installed) `v2.4.2`,
   resulting in a hash mismatch (because of the patch change) when running `yarn install --immutable`

The fix:

- Upgrade yarn in development environment and run `yarn install` again
- Change `./setup.sh` to require a specific version of `yarn` (currently `2.4.2`)

This is only expected to fix the `yarn install` portion of CI (not typechecking yet)

+++ setup.sh
@@ -11,6 +11,9 @@
 
 SPLITGRAPH_DIR=${1-"$(cd -P -- "$(dirname -- "$0")" && pwd -P)"}
 
+# If `yarn --version` is not exact match of `$TARGET_YARN_VERSION`, install yarn
+TARGET_YARN_VERSION="${TARGET_YARN_VERSION-"2.4.2"}"
+
 prep_env() {
     echo "Ensure certs..." \
         && echo "Ensure yarn..." \
@@ -157,6 +160,11 @@ dir_has_yarn_release() {
         rm -f "$prefixDir"/.yarn/plugins/@yarnpkg/plugin-constraints.cjs
     fi
 
+    if has_wrong_yarn_version "$prefixDir" ; then
+        echo "yarn seems outdated in $prefixDir, return 1 to trick into upgrade"
+        return 1
+    fi
+
     if has_broken_yarn "$prefixDir" ; then
         echo "yarn seems broken in $prefixDir, remove .yarnrc.yml"
         rm -f "$prefixDir"/.yarnrc.yml
@@ -191,6 +199,32 @@ dir_has_yarn_plugins() {
     return 1
 }
 
+
+# OPTIONAL:
+#   If $TARGET_YARN_VERSION is set, then `yarn --version` must be exact match
+has_wrong_yarn_version() {
+
+    # specifying a target version is optional, so bail out if not specified
+    if test -z "$TARGET_YARN_VERSION" ; then
+        return 1
+    fi
+
+    local prefixDir="$1"
+    shift
+
+    pushd "$prefixDir"
+
+    local yarn_version
+    yarn_version="$(yarn --version)"
+
+    # For now we just check exact match
+    if test "$yarn_version" != "$TARGET_YARN_VERSION" ; then
+        echo "yarn --version ($yarn_version) is not TARGET_YARN_VERSION ($TARGET_YARN_VERSION)"
+        popd && return 0
+    fi
+
+    popd && return 1
+}
 
 # Sometimes a mismatch between .yarnrc.yml yarnPath and the actual yarn file
 # can cause yarn to fail to start due to looking for the wrong file
